+++ packages/manager/src/containers/withLinodes.container.ts
@@ -1,4 +1,3 @@
-import { StackScript } from '@linode/api-v4';
 import {
   CreateLinodeRequest,
   Linode,
@@ -12,16 +11,6 @@ import {
   useCloneLinodeMutation,
   useCreateLinodeMutation,
 } from 'src/queries/linodes/linodes';
-import { useStackScriptsOCA } from 'src/queries/stackscripts';
-import { baseApps } from 'src/features/StackScripts/stackScriptUtils';
-import { useFlags } from 'src/hooks/useFlags';
-
-const trimOneClickFromLabel = (script: StackScript) => {
-  return {
-    ...script,
-    label: script.label.replace('One-Click', ''),
-  };
-};
 
 interface Actions {
   cloneLinode: (data: {
@@ -59,37 +48,5 @@ export const withLinodes = <Props>(
     linodesData,
     linodesError,
     linodesLoading,
-  });
-};
-
-export interface WithMarketplaceAppsProps {
-  appInstances: StackScript[] | undefined;
-  appInstancesError: string | undefined;
-  appInstancesLoading: boolean;
-}
-
-export const withMarketplaceApps = <Props>(
-  Component: React.ComponentType<Props & WithMarketplaceAppsProps>,
-  enabled = true
-) => (props: Props) => {
-  const { data, error, isLoading } = useStackScriptsOCA(enabled);
-  const flags = useFlags();
-  const newApps = flags.oneClickApps || [];
-  const allowedApps = Object.keys({ ...baseApps, ...newApps });
-
-  const filteredApps = (data ?? []).filter((script) => {
-    return (
-      !script.label.match(/helpers/i) && allowedApps.includes(String(script.id))
-    );
-  });
-  const trimmedApps = filteredApps.map((stackscript) =>
-    trimOneClickFromLabel(stackscript)
-  );
-
-  return React.createElement(Component, {
-    ...props,
-    appInstances: trimmedApps,
-    appInstancesError: error?.[0]?.reason,
-    appInstancesLoading: isLoading,
   });
 };
